Changed the package name from de.gnmyt.SQLToolkit to de.gnmyt.sqltoolkit
