+++ src/main/java/de/gnmyt/sqltoolkit/factory/TableFactory.java
@@ -1,8 +1,8 @@
-package de.gnmyt.SQLToolkit.factory;
+package de.gnmyt.sqltoolkit.factory;
 
-import de.gnmyt.SQLToolkit.drivers.MySQLConnection;
-import de.gnmyt.SQLToolkit.storage.SQLStorageMedium;
-import de.gnmyt.SQLToolkit.storage.SQLTable;
+import de.gnmyt.sqltoolkit.drivers.MySQLConnection;
+import de.gnmyt.sqltoolkit.storage.SQLStorageMedium;
+import de.gnmyt.sqltoolkit.storage.SQLTable;
 
 import java.util.HashMap;
