+++ src/main/java/de/gnmyt/sqltoolkit/types/TableField.java
@@ -1,4 +1,4 @@
-package de.gnmyt.SQLToolkit.types;
+package de.gnmyt.sqltoolkit.types;
 
 public class TableField {
